#75. Moved ProjectsApi contracts methods to ContractsApi.
